♻️ refactor: make agent tracing core, not a plugin
Move the trace hook span logic into internal/observability/tracehook
and append it unconditionally in agent hook construction, removing the
disableable hook/trace plugin. Drop the stale plugin row at startup.
