+++ cmd/stella/plugins_imports.go
@@ -20,7 +20,6 @@ import (
 
 	// Plugin hooks.
 	_ "github.com/CherryHQ/stella/plugins/hooks/rtk"
-	_ "github.com/CherryHQ/stella/plugins/hooks/trace"
 
 	// Plugin sandbox backends.
 	_ "github.com/CherryHQ/stella/plugins/sandbox"
